wookbook

+++ app/src/main/java/com/kaka/gg/Login.java
@@ -7,6 +7,7 @@ import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.text.TextUtils;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -21,7 +22,6 @@ public class Login extends AppCompatActivity {
     EditText id, pw;
     SharedPreferences shard;
     SharedPreferences.Editor editor;
-    Boolean loginChecked;
     String res = "";
     String cid, cpw;
     Lc lc;
@@ -46,11 +46,12 @@ public class Login extends AppCompatActivity {
                 if (res.equals("Su")) {
                     id.setText("");
                     pw.setText("");
-                    Toast.makeText(Login.this, shard.getString("name", "") + "님 자동 로그인성공", Toast.LENGTH_SHORT).show();
+                    Toast.makeText(Login.this, shard.getString("name", "") + "님 자동로그인 성공", Toast.LENGTH_SHORT).show();
                     Intent i = new Intent(Login.this, MainActivity.class);
+                    i.putExtra("newid", shard.getString("name", ""));
                     startActivity(i);
                 } else {
-                    Toast.makeText(Login.this, "최초 로그인을 해주십시오", Toast.LENGTH_SHORT).show();
+                    Toast.makeText(Login.this, "로그인을 한 번 이상 해주십시오", Toast.LENGTH_SHORT).show();
                 }
             }
         });
@@ -58,9 +59,16 @@ public class Login extends AppCompatActivity {
         autostop.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                editor = shard.edit();
-                editor.clear();
-                editor.commit();
+                lc.loch(shard.getString("id", ""), (shard.getString("pw", "")));
+                res = lc.gets();
+                if (res.equals("fail")) {
+                    Toast.makeText(Login.this, "해제할 아이디가 없습니다", Toast.LENGTH_SHORT).show();
+                } else {
+                    editor = shard.edit();
+                    editor.clear();
+                    editor.commit();
+                    Toast.makeText(Login.this, lc.Name.get(lc.k).toString().trim() + "님 자동로그인 해제", Toast.LENGTH_SHORT).show();
+                }
             }
         });
 
@@ -88,9 +96,12 @@ public class Login extends AppCompatActivity {
             editor.putString("id", cid);
             editor.putString("pw", cpw);
             editor.putString("name", lc.Name.get(lc.k).toString().trim());
+            String newi = lc.Name.get(lc.k).toString().trim();
             editor.commit();
             Toast.makeText(Login.this, lc.Name.get(lc.k).toString().trim() + "님 반갑습니다", Toast.LENGTH_SHORT).show();
             Intent i = new Intent(Login.this, MainActivity.class);
+            i.putExtra("newid", newi);
+            Log.e("newid", newi);
             startActivity(i);
         } else if (TextUtils.isEmpty(cid) || TextUtils.isEmpty(cpw)) {
             Toast.makeText(Login.this, "ID또는 PW를 입력하십시오", Toast.LENGTH_SHORT).show();
